feat: change app layout
