+++ widgets/AppLayout/ui/GradientBackground.tsx
